VCL: move RB Accept header normalization to text-fe
Perform RESTBase Accept header normalization on the frontends only. Path
normalization needs to be done on both frontends and backends for PURGE
requests to match stored paths. Varnish removes all variants defined by
Vary when purging, so there is no need to deal with headers at the
varnish-be layer for purging purposes.

Ref: https://varnish-cache.org/docs/5.1/users-guide/purging.html#http-purging

Change-Id: If2681a183445cf51b15d799306b0e66bd9570c26