Monitor dump output file production
We've had a couple instances in the past where dumps hang until
they are killed (bad MW deploy). It's rare but this job will
notify us if it happens again, within a few hours of there
being an issue.

This runs only on the active NFS server for xml/sql dumps.

Change-Id: I367a1008b6293501e5e8175a681e6dbd7849291c