Remove at
When looking through further candidates for automatic service restarts
I noticed that atd is only installed on trusty/jessie systems. Turns
out that starting with stretch the priority got lowered to 'optional' so
it's no longer installed in d-i base installation.

We now have more stretch than jessie hosts, so it seems fairly obvious
that noone in SRE is using atd, so remove it. It's one less service
running which can potentially cause issues.

I also checked reverde dependencies and there are none (devscripts
as installed on boron recommends it and a few older kernels have
a Breaks:)

Change-Id: Iae7707006976bdd4a408ce9e38e23d6dc6a2d3b8