ircecho: remove restart => true parameter, unbreak
base::service_unit does not have a parameter named "restart".

systemd::service does.

Ie0881fdaba78d6 added it to systemd::service

Icdf33762efd33cb65 was supposed to be a revert of 606e47f134ea3c
but it did not remove that parameter from the separate change.

This lead to puppet error on einsteinium:

Change-Id: Ib30ece38bf7ae1ee7b2ab338d0d4c80908ebfb91