profile::hadoop::common: selectively deploy Ores packages
While adding the package hunspell-bs in ores::base a lot of puppet failures
popped up since it is available only in Stretch, and due to
profile::hadoop::common a lot of jessie analytics hosts have those
packages too. This change limits the deployment of ores::base packages
only to the Hadoop worker nodes and the stat machines.

Change-Id: I7360a272d7cbcb2fd377acac8ee1270b6592329f