icinga: Populate additional hostgroups based on LLDP
Use the lldp_parent fact we provide to add one extra hostgroup to all
physical hosts per the network switch they are connected to. Filter
based solely on the name of the peer, which is fragile but should work
for now.
This is a preliminary approach that only targets hosts connected to
switches. It can probably be augmented

Bug: T187962
Change-Id: I782761c70ccbc5bc2999c9f14c14adeebb7879ef