phabricator: List new and recent assignees
List all Phab user account names who registered their user account in
the last six weeks and who are a task assignee and one of the tasks
they are assigned to got changed somehow in the last week.

This allows to potentially follow up with new developers, in addition to
experiments how to better track new developer activity in Gerrit (T188244).

Bug: T195780
Change-Id: I95df00af4c7d069fb6e12b12262da672a893f5fa