utils: add script to generate mcrouter-related certs
We use cergen under the hood to generate the certificates for all of the
existing servers with the mediawiki class installed according to
puppetdb.

The advantage over a simple manifest file for cergen is that we can
dynamically generate the certs; the disadvantage is we need a first
puppet run (which will fail to start mcrouter) before the certs get created.

Bug: T192771
Change-Id: Id308ae581efb452085778b66e8b6d7463a300068