deployment-prep: Remove override for scap::sources
This should not need to be duplicated between prod and beta,
the default from hieradata/role/common/deployment_server.yaml
is appropiate for both prod and beta.

This was previously causing role::webperf to be broken in Beta,
because statsv/statsv is defined deployment_server.yaml but not
repeated in the copy for deployment-prep/common.yaml. Rather
that pointlessly trying to keep these in sync, just use the same
for both.

Bug: T195314
Bug: T196034
Change-Id: Ie7aad8939691faa06a63b9bf163bd347b5cfdd65