webperf: Add navtiming and coal to scap::sources
In order to bootstrap the service on a target (e.g. webperf1001),
it needs to fetch from the deployment server (e.g. tin).

The scap::target() resource used by webperf/navtiming attempts to
do this when no local copy exists at /srv/deployment, but this
failed in Beta Cluster because the scap source (beta's "tin")
also didn't have it.

This was manually created on tin, but should've been done by puppet
using the scap_source function, which is automatically called for
each entry in scap::sources on a deployment server (such as "tin").
That way things won't break when production switches to deploy1001.eqiad
or naos.codfw as primary, and also makes things work in Beta.

The scap_source() function essentially does the following two steps
from the Scap3 manual:
- Ensure git checkout exists in /srv/deployment.
- Run `scap deploy --init` once, if it was never run before.

Update scap_source() documentation to reflect the fact that
the default repository is prefixed with mediawiki/services/,
which was rather counter-intuitive.

Bug: T195314
Change-Id: I513a5a225ae4e4a8ff6223f37aa69f0cbf0a1366