Prepare to tighten Puppet DB access control - check client certificates
Currently profile::puppetdb does a simple IP restriction with ferm.

In T194962, the ability to send a command to Puppet DB to add a resource to a
host will become more security-sensitive, so let's lock that down.

The puppetmasters already send client certificates, this just starts checking them.

Right now we don't take action based on the verification results, we just log them.

Bug: T194962
Change-Id: I692226d37d9a183e8c3c3ddd1830bae2188bccaf