varnish: Remove setting of CP cookies
Introduced in 2015 with 4c6a29bb3c, for the purposes of determining
HTTP/2 usage client-side for EventLogging via the NavigationTiming
extension. That data was subsequently going to be used to inform
decisions about T119797 and other things that could be motivated
for or against based on how common HTTP2 is.

This data has not been used for some time now, and instrumentation was
removed in Feb 2018 (b3eccc716; T186295). I also could not find any
undocumented users of it since then, via CodeSearch.

Bug: T110353
Change-Id: I69b81d29007d7364a9f8b5b1a333add8fd6d2284