openstack: allow designate in labtest to contact labtestn keystone
labs-ip-alias-dump is failing here I believe because it looking
towards the incorrect host for keystone and if it did try to
contact the correct host it would fail due to ferm. This commit
separates the idea of nova_controller and keystone_host for
base/pdns and passes in the correct value while opening up
iptables.

Note labs-ip-alias-dump calls this nova_api_url which is a
confusing name and should be addressed in another commit.

Bug: T167559
Change-Id: I4bd8c702f1137f77e5dc98ad22509a8005065830