phabricator: set service IPs for phab1002 in Hiera
Set the listen address (phab1002.eqiad.wmnet) and
the service IPs (phab1002-vcs.eqiad.wmnet) for phab1002.

Compare to ./hosts/phab1001.yaml.

Note how there are 2 remaining lines commented out. This
is for the "unmapped" IPv6 address that we don't have on phab1002
and git-ssh.eqiad.wikimedia.org which will have to be switched
over by removing it from phab1001 and adding it on phab1002.

Bug: T196019
Change-Id: I46b787c686f642aef1bf16149e3fef3899785952