mw-maintenance: require GNU time from time package
On terbium the 'time' package is installed, providing the
GNU version of time.

It provides more info than the shell built-in time command.

Anomie pointed out he needs it and that it was installed
on tin but isn't installed on mwmaint1001 on stretch.

Thus, requiring it now in the mw-maintenance profile.

Bug: T192092
Change-Id: Idaab8cd367b2d49ec38954f3f298337eb3d60f1e