role::aqs: deploy different cassandra versions
After the upgrade of aqs1004 to cassandra 2.2.6-wmf5
we discovered that the new dependency, python-cassandra,
on jessie causes cqlsh to fail for a import error. This
is probably due to the mismatch between the version
available via jessie backports (2.5) vs stretch (3.7).

Since there is a plan to reimage all the aqs nodes to
Debian Stretch in the near future, it is probably not
worth it to find a solution for Jessie but only keep
this workaround in place to please puppet.

Bug: T197062
Change-Id: Ibd1b530a41782ac6d9175974bfffa7b59645323d