icinga/wikidata: add secondary dispatcher critical check
We switched the normal wikidata dispatcher check to be just
a WARN to watch it for some time. That warns if over 10m,
here we are now adding a second check that is the same
thing but returns a CRIT if it's 4000s or more.

This is based on "anything under an hour is not really
critical" while we still want to be informed if for some
reason the queue is fully stopped.

Change-Id: I76c1f27a8764baa08436707c0241b40114fe40ee