role::logstash: fix gelf filtering
In Iba94cf067401260,  the GELF filtering was corrected to use
full_message instead of short_message, if available; unfortunately,
while this is needed for elasticsearch (where short_message is often
truncated) it is actually harmful for any nodejs service that uses
gelf-stream, which json-encodes data in full_message.

So, maintain the best of both worlds if possible, and be compatible with
both formats.

Bug: T197219
Change-Id: I9f70a15eee06882d918a345f1ebd322d212598b1