cassandra: add another package version to the 2.2 list
The 2.2.6-wmf5 package has been tested on Debian Stretch
and it works nicely, but it gives some problems with Jessie
as described in T197062. The simplest solution is, in my opinion,
to add another use case to the cassandra package list, and wait for
the upcoming reimage to Stretch of all the Cassandra 2.2 clusters
(currently only AQS, but maps will jump to 2.2 on Stretch directly).

Bug: T197062
Change-Id: I39590a96992efb9f07773fbbbc2b24b710291ed2