webperf: Get graphite_host for coal::processor from Hiera
This was hardcoded to eqiad, which could be bad in case of a switchover,
but is also currently causing the service to not work in Beta Cluster.

There already exists a field for this in Hiera (graphite_host),
and it's already set universally for Beta Cluster.

For production, there is a convention to set this at the role level,
rather than globally. I assume this is because we have multiple graphite
masters in prod, and different services are tied to different ones. And
it seems these are not centrally defined or aliased in some canonical way
within Hiera. This means we'll have to follow the same pattern as for
AQS, Restbase, Maps and other services and define it in Hiera for prod's
webperf::processors_and_site role.

Bug: T195314
Change-Id: I7ac009b644ce2d2ee37b42b041de89f4833f600b