phabricator: Use the mysql native driver
Switch to using the mysql native driver in phabricator. This was
suggested by the phabricator software itself when logged in as an admin
where an "Unresolved Setup Issues" with a yellow icon would show up and
who say that MySQL Native Driver is Not Available. Note that
php5-mysqlnd is conflicting with php5-mysql. That is the 2 can not
coexist as they offer the same functionality.

Change-Id: I1b6bf28cba6f044189943c8f5b64430c7ba072c7