graphite: start sending metrics to graphite2003 too
This will allow the host to have recent data, the rest of datapoints will be
merged in the background from graphite200[12]. Starting next week graphite2003
can be switched to be the only codfw graphite host.

Note that we are effectively merging "big_users" and "default" clusters since
Cassandra metrics now have moved to Prometheus.

Bug: T196483
Change-Id: I9fa7518e6fe93fc93409cc7aa578fccbc6515311