openstack: add openstack-cvps wrapper script
This script is a simple wrapper for the openstack command that
reads the environments variables (novaenv) before launching the actual command.
It allows you to interact with openstack using sudo rather than being root all
the time.

The 'cvps' prefix stands for 'cloud vps'.

Example usage:

 user@machine:~$ sudo openstack-cvps service list

Previous to this wrapper, one would need:

 user@machine:~$ sudo su
 root@machine:/home/user/ $ cd
 root@machine:~# source ./novanenv.sh
 root@machine:~# openstack service list

Bug: T196633
Change-Id: Id5475ea297a109f0b097fd088676bd916b495a9b
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>