profile::mariadb::misc:el::sanitization: move whitelist to Refinery
The Analytics team moved the Eventlogging purging whitelist to
the Refinery, and hence this patch takes care of changing the puppet
dependencies according to this new setting. This implies deploying
the Analytic Refinery on db110[7,8].

To avoid deploying too many unrelated packages to the database hosts
(and potentially other use cases like this one), I created the
profile::analytics::refinery::repository profile.

Bug: T198766
Change-Id: Ic73d9134b60c11221679e9a723c99c1cbdbc1edd