Enable base::service_auto_restart for SSH
Automatically restart the SSH daemon if underlying runtime library
dependencies have been upgraded.

This may sound scarier than it is (as it would also affect Cumin)
but it's worth keeping in mind:
- OpenSSH restarts are very reliable, I never ran into a restart
errors for the all the past restarts over the recent years
- We don't upgrade software automatically, so when e.g. a new
reverse dependency is rolled out, restarts of base services are
tested anyway. If those tests fail, the broken library would not
get deployed in that state.
- wmf_auto_upgrades are splayed over the day (and only from Monday
to Friday), so if something would still manage to bypass the
earlier checks, we'd notice it via the Icinga check which detects
broken systemd service states.

This also been tested with wmf-auto-restart for a number of different
setups (dns4001, ores1001, mw1318, analytics1003, ms-be1028, neon,
ms-fe2008, wtp1026, kafka-jumbo1006).

I'll also call this out in the SRE meeting to allow anyone to object
if deemed too risky.

Bug: T135991
Change-Id: Iedc1a650042289e5d341a8f94f6ee1b56d56d794