Blacklist floppy driver
I'm doing a review of past kernel security issues and whether they affected
functionality which is otherwise unused by us to minimise our attack foot
print.

The floppy driver was affected by a vulnerability in one of the ioctls
provided by the driver which resulted in an information leak.

It turns out floppies are still a thing! Not in our physical servers,
but emulated by Qemu in both the Ganeti cluster and in OpenStack.
As we don't really need floppies, let's blacklist it to prevent
future/unknown vulnerabilities in the floppy driver.

Independant of this change it's worth investigating how to disable floppies
in Ganeti/Cloud VPS for future instances (or retroactively for old ones).

Change-Id: I9a9f0f8d240f10339d299e9c245a8b8e3a74db68