webperf: Split Redis from the rest of the arclamp profile
Rename the Xenon module to Arc Lamp. The "Xenon" name was confusing because
it was also the name of the HHVM component where the data originates on
MediaWiki servers.

In 2015, Ori already used the name Arc Lamp to publish this puppet module in
its repository at <https://github.com/wikimedia/arc-lamp>, but the Puppet
classes retained their original name (until now).

This commit is in preparation for two changes:

1. Generalise the infrastructure for receiving stack traces from PHP 7,
   in addition to HHVM/Xenon.

2. Move processing and HTTP-hosting of stack traces from mwlog1001
   to the webperf1002 node. But keep the Redis instance on mwlog1001
   given that it will be written to from online production traffic.

Bug: T195312
Change-Id: Ie9f9020c8bcfde7bb96791a3cfd74279eae668a8