profile::mediawiki::web_testing: create profile
We move apache-fast-test to its own profile, together with some specific
urls it can use for doing smoke tests on the various virtualhosts in the
WMF MediaWiki apache configuration. These are good mostly for verifying
the resolution order is unchanged, more specific apache configuration
changes will need more accurate tests.

Also, install the profile on the cumin masters, as apache-fast-test can
be useful to use in scripts running on the cumin masters.

Change-Id: I70e5ce39f41fe717d26c5b651b92a6510f8188bf