varnish: startup process for multiple VCL files
Start varnishd(1) without specifying a VCL file (-f ''). Use reload-vcl
in ExecStartPost to load and label separate VCL files first, and the
main one last. This is necessary given that the main VCL file might
reference the label of separate VCLs, and would otherwise fail to
compile if loaded directly with -f.

Further, varnishd refuses to start the child process if called with an
empty -f. Make reload-vcl do that with a new argument, --start-child.

Bug: T164609
Change-Id: I0206915ec44aa5286b37d8585224e51c292e6c48