network::constants: define all caches, not only cache_misc
As a prerequisite to merge cache_misc into cache_text, we need to ensure
that all cache_text hosts can talk to the misc origin servers. More in
general, there is no real benefit in limiting access to the origin
servers to a subset of the caches, and allowing them all will make
future reorganizations easier.

Bug: T164609
Change-Id: I5743f32f9a3aa779081c65263bae2ebb9977f1f9