Delete unused code in elasticsearch module
The proxy stuff looks to be dead code, un-referenced since 2016 when
nobelium was decomissioned as an experimental elasticsearch replica
for labs.  The nginx instance running on the prod search clusters all
go through elasticsearch::https class instead of this one.

The diamond parts were all in an ensure => absent stage and hadn't
been cleaned up yet.

Change-Id: I481d16549becab603328be3ba46f746de6bd357d