phabricator: Print IDs of projects of tasks assigned to disabled accounts
When a Phab user account (e.g. former WMF employee) gets disabled we
reset the assignee field of tasks still assigned to such an account
to avoid cookie-licking.

Several times in the past, even a Phabricator admin could not find
all tasks assigned to such accounts (as Phabricator admins are not
all powerful, see e.g. mw:Phabricator/Permissions#Administrators),
for example because the task is in a restricted Space.
In such cases, the project(s) or Space associated to those tasks
need to be assumed and then another person needs to be found that
can still access these tasks in order to reset the assignee. This
has required contacting people just to find out that assumptions
were wrong. This is wasting time, hence at least print the project
PHID(s) of tasks associated to that Phab user account (cheaper
than querying the actual name which requires another JOIN). The
associated project name can be found via Conduit's project.query.

Change-Id: I0a295cc474fcecce411d9f5366bcad7e3aac415c