labtestn: make the services node more like a normal services node
I'm still thinking this through but I think it's fine if
this is the same as the labtest services node.

We may or may not share the same DB but otherwise we may
as well run a full DNS setup here; this is a model for
labservices1003/1004 which will eventually become the primary
services/DNS servers for eqiad1.

Change-Id: Ie179a7b1112c2819ac24e0f1ba80233b1f59c072