EventStreams: Consume messages from the local Kafka cluster
EventStreams in CODFW has been consuming enourmous amounts of memory,
partially because it is connecting to the Kafka cluster in EQIAD. Since
in that DC there are no such problems, temporarily have ES in each
cluster consume messages from the Kafka cluster local to the DC the
service is in. While clients will still be perfectly able to consume
messages, their offsets will be reset by doing this.

Bug: T199813
Change-Id: I2c083b215c39516811d9193f579bdcd97df6dc7b