allow storage generated misc cron dump output on secondary nfs server
This introduces new roles and profiles so that misc cron dumps
can easily be split off from the NFS server used to store the
xml/sql dumps.

The actual move will be done in a followup patchset.

Rsyncs will go from
  xml/sql dumps nfs server to its secondary and to public-facing servers
     -- all dumps, all status files
  misc dumps nfs server to public-facing servers
     -- all dumps
  misc dumps nfs server to xml/sql secondary
     -- adds/changes and category dumps only
     No other dumps require data from a previous run, so failover
     won't be an issue for them

We keep around an extra profile so that it's relatively easy to switch
everything back to one host when needed; hiera switches are added
so that generating hosts know which backend to use for the nfs mount;
templates and manifests are renamed from 'primary' and 'fallback' to
names indicating the specific usage (xmldumps, misccron, all).

Also: fixed up role comment strings, some of them were obsolete.

Bug: T200180
Change-Id: I0aa228a0645aa2d87380329b7e894949fe09d936