wiki replicas: moving compatibility views to $table_compat
This is a "have your cake and eat it, too" option.  The baseline is
using the tables roughly like they are with filtering,
but for tools where this breaks things and refactoring is too hard
it provides the option of using "_compat" views where older code should
continue working, albeit more slowly.  The same notion could be applied
to the comment table and for backward compatibility on the actor table.

This more or less removes the penalty for those keeping code up to date
while being somewhat easier on those who can refactor only as
little as possible to keep something running.  It'll still require
some refactoring, but the penalty hits the older code rather than
well-maintained code.

Bug: T174047
Change-Id: I70b92738c9da94748c64c3ac99b5c73ae94c8c6c