planet: drop feed templates for planet-venus/jessie
For a transitional period we supported different
software on jessie (planet-venus) and stretch (rawdog)
to create the Planet feeds.

This meant we had to support 2 slightly different formats
of feed templates.

This is dropping the support for planet-venus on jessie by removing
those.

We only use stretch now and the upside is people don't have to make
edits in 2 locations to add or remove a single feed.

Bug: T180498
Change-Id: Ia30f7651dd03063a41136820a962bd4833dde2e1