storage config tweaks for cp1075-99
The Samsung pm1725a nvme drive defaults to a lower-performance
compatibility LBA mode that represents itself in terms of legacy
512-byte blocks.  The "nvme" command from nvme-cli can switch this
to native 4K blocks that the kernel can optimize better for, but
this must happen before partitioning, so all work on the nvme
drive is done in late_command rather than partman.

The mkfs parameters are also tweaked significantly from previous
efforts, we'll see in practice if we need to back any of this out.
Our use-case is a giant filesystem with 1-5 files filling it, and
no requirement that the storage is reliable across partial disk
failure or host crash (we can always re-mkfs and repopulate the
cache from scratch).  The bigalloc feature is turned on, using
16MB allocation chunks.  The journal is removed completely, and
all other metadata is minimized and shoved to the front of the
disk with no superblock backups.

TODO in later commit for the runtime puppetization:
* support /srv/cache and single partition, etc...

Bug: T195923
Change-Id: I999e91656ff5f8df54ad0c2fb54f1a921c12bc78