osm: Populate the osmupdater user
The osm labs powering puppet code has gone through a number of back and
forths with the maps powering puppet code. While there is some code
sharing the 2 infrastructures differ and the code ends up diverging in
places causing issues. This change is reallinging once more the
osmupdater user. This is not tenable in the long run and probably the
labs related osm profile/role and infrastructure should be deprecated
and removed

Bug: T197246
Change-Id: I0cc9035010a6c4d2c7b7a6bd43e40f564bd58e29