cp1075-90: define in site.pp
Note odd numbers are text, even are upload.  This is just how the rack
layout works out for redundancy.  There are 16 total nodes, and there
are 4 hardware rows.  2 nodes are placed in each of 2x separate 10G
racks in each row.

This means for a given 8-node cluster (text, upload), there are 2 nodes
in each of the 4 distinct rows, and those two nodes are in distinct
racks within the row, connected to distinct per-rack switches.

Bug: T195923
Change-Id: Ib61680f1d2bd0da0edb411ba88132d5d1f353a79