varnish: drop -Werror from cc_command
By default, cc_command is defined as follows:

  exec gcc -std=gnu99 -g -O2 -fstack-protector-strong  -Wformat \
        -Werror=format-security  -Wall  -Werror \
            -Wno-error=unused-result -pthread -fpic -shared -Wl,-x -o %o %s

We do override cc_command if geoip is enabled, omitting most of the
flags above.

If geoip is disabled, we use the default cc_command. On stretch, VCL
with inline C fails compiling with errors like this in that case:

  vgc.c:2449:39: error: ‘vgc_probe_varnish’ defined but not used [-Werror=unused-const-variable=]
   static const struct vrt_backend_probe vgc_probe_varnish = {

Set cc_command to the default regardless of whether geoip is enabled or
not, but drop "-Werror -Wno-error=unused-result".

Bug: T200445
Change-Id: Id78fdd6a20e318c3c281ecd0462d524f1c1ae52a