mw::maint: allow home dir rsync between mwmaint, not terbium
terbium is not used anymore for mediawiki maintenance and has
a decom motd.

User home dirs have been synced from there to mwmaint1001.

So now update that to allow rsyncing between mwmaint1001 and
mwmaint2001 and remove terbium.

It is not set to automatically sync, the command still needs to be
run manually but it prepares the rsyncd and ferm config to allow it
once it will be needed to switch over to codfw.

Bug: T192092
Change-Id: I9e75f88303698752bf920b375070658881953dcd