ip -6 route show, don't try to find the IP, but keyword instead
Follow up from https://gerrit.wikimedia.org/r/c/operations/puppet/+/437784

`ip -6 route show 2620::860:104:10:192:48:30/128 | grep -q 2620::860:104:10:192:48:30`
Is used to check if a route is already installed,
but `ip -6 route` displays that same IP as "2620:0:860:104:10:192:48:30".
(Replaces "::" with ":0:")
 Which makes puppet think it needs to re-add the route, which fails.

As all the routes we add contain the keyword "via",
we check for the presence of that keyword of the `ip route` output.

Bug: T195365
Change-Id: Ib232a441d73f569c3f78554525f97d63bfc40dbc