db backup statistics: Initial implementation of the backup stats
At backup start time, store the ongoing nature of the new backup,
and, if it fails, or it is successful, update it with the
corresponding state on the zarcillo database.

While at the moment this only monitors the existance of the
backups, it is an initial statistic to check backups finish
correctly, with the final goal of more fine-grained statistics
not yet implemented.

Bug: T198987
Change-Id: Ie8a9e64a9d9d6a668ef79835021cf3a401e8cec6