mariadb: Set pages for multi-instance hosts
The idea is to be able to page on a instance level for the multi-instance core servers.
Right now it is not possible to page only on a given instance on a given host.
We have multi-instance hosts where we don't want to be page and it is enough
with an IRC alert - ie: misc hosts.

However, for core multiinstance we do want a page, as long as it is on the active DC.

This change is based on the implementation of the "auto" flag (449742) to auto-detect if the
host belongs to the primary DC (while we are still on an active-active) configuration
or not. And if it does, page.

Bug: T200509
Change-Id: I9e99d3e197336732d4c111be2f309540ddc584d4