PHP: create module for modern Debian-based distributions
This module should be used to set up and install php 7.x on debian
distributions. This first commit only manages the base configuration of
the various sapis and of the relative extensions. Services are not
managed, specifically php-fpm, support for which will be added in a
subsequent patch.

Also add a metadata.json file to the wmflib module so that we can
explicitly declare the dependencies in this module.

Bug: T201140
Change-Id: I7c76108cc00d03aff4e2a87633748399a78c4424