mariadb: Make wikireplicas obey the current mariadb read_only config
I think this was intended to be 1, based on:

read_only     => 'ON',

config from mariadb::config. Changing it manually, and checking
with cloud in case they can see if there was a reason for this to be
missconfigured. Even if this was correct and supposed to be 1 (e.g.
due to some script not having SUPER), this was still a confusing
bug.

Bug: T172489
Change-Id: I5b05cb2e149bd892af55ef54503714016392d2e3