remove old eqiad caches from data lists
The hieradata change will get them out of ipsec sets, and the
conftool out of pooling data.  Need downtimes and puppet disables
before this, and reimage of the 19 affected systems here (14 to
decom, 4 spare for ATS, 1 spare for canary).  Next commit shows 18
as decom because it includes the old ex-cache_maps set as well.

Change-Id: I51c92de217d2ab0a81fc2bc8884c0f856f066465