netbox: make the role usable on a stand-alone host again
There was first a standalone role for netbox and then it
was included in the netmon role which has profiles for netbox
and other applications.

As part of If976912e3cb01d208 it was converted to use the
httpd class rather than apache.

This meant the declaration of the httpd moved to role(netmon).

Trying to use role(netbox) standalone on a cloud VPS to test things
resulted now in a dependency error. The letsencrypt setup was
missing the apache2 resource.

To make this role usable by itself again while also keeping the
puppet code nice add a httpd for netbox profile and include it in the role.

Production isn't affected because it just uses the netmon role which
in turn includes httpd and the netbox profile.

The modules neeed can be compared on
https://gerrit.wikimedia.org/r/#/c/operations/puppet/+/409210/4/modules/profile/manifests/netbox.pp
where the former apache::mod:: classes were used in the past.

Change-Id: I32f574baad8371f62e42201ad7b18a8ac3af79c1