netbox: don't hardcode db_master, use active_server from Hiera
Don't hardcode the hostname netmon1002 as the db master to
make the role usable on a cloud VPS for testing. (for T185504)

Not adding it to Hiera because it seemed like a duplicate
of the existing $active_server.

Afaict the active_server would also always be the db master so
just setting it to that.

Change-Id: I039434602bc578ad8465357a06478ad885d96d69