webperf: Switch webperf::site to use arclamp from webperf-2
The previous commits prepare profile::webperf::arclamp to work
on both mwlogX001 and webperfX002, and to be multi-dc friendly,
and to work on both webperf1002 and webperf2002 (in both DCs).

Once that is merged, this commit will switch webperf::site (runs
active-active on webperf1001 and webperf2001) to proxy arclamp
data from webperf1002 and webperf2002, instead of currently where
it always reads from mwlog1001 in eqiad.

Bug: T195312
Change-Id: I35899ad4c4c2c4ad84d23180b26381d6d51ec5c4