releases/mediawiki: proper Icinga monitoring for both Apache vhosts
Nowadays a releases::mediawiki server has 2 virtual hosts, the regular
releases host and the releases-jenkins host.

This unspecific http check happened to hit releases-jenkins first,
resulting in an alert on releases2001 where jenkins is currently not
running and which isn't configured as a backend for it in varnish either.

Have sepearate checks for the seperate virtual hosts to make it obvious
where the issue is (jenkins or unrelated).

Change-Id: I40409ddf33085aeeddd467ed8dcfe9d6f1b165eb