Enable intel-microcode for all bare metal servers with an Intel CPU
Microcode has been gradually enabled across the fleet via a Hiera
flag and at this point it's enabled for 926 hosts out of our 1221 baremetal
servers. No problems have been encountered and at this point it seems safe
to install it by default.

On virtual instances we don't need microcode updates, it is being passed
through by KVM from the virtualisation host.

(The Hiera framework to enable it selectively will be removed in a followup
commit).

Bug: T127825
Change-Id: Ib8cb0e6dad25b0c54b7d70549fd73b44eb51eac5