db-master.cfg: Enable haproxy health-check logging
We found out that dbproxies were not logging status changes for the
healtchecks, so it was impossible to know when a failover was triggered.

As per the documentation, "option log-health-checks" "option tcplog"
"log /dev/log local0 info" looks like what we'd need:

"When this option is enabled, any change of the health check status or to
the server's health will be logged, so that it becomes possible to know
that a server was failing occasional checks before crashing, or exactly when
it failed to respond a valid HTTP status, then when the port started to
reject connections, then when the server stopped responding at all."

I would ideally test this on a proxy that is not in use (like
dbproxy1006) before deploying this everywhere.

Bug: T201021
Change-Id: I62d1590f7840fa51acbfe3c5f9979720d4276c0b