keystone: Limiting worker process numbers
By default keystone will make one worker process per processor on your
system when not using apache for service.  While we probably should
change to apache in the future (as this appears to be best practice),
while using the eventlet server directly, we need these values to stop
it from making nearly 100 processes, each opening its own DB connections.
I think this was the root cause of the problem in the bug below.

Bug: T188589
Change-Id: I4ad5ae3c85d33e0f18820aca175eeb24bab80fbe