nova: limit metadata workers
Since the number of metadata workers also equals the number of CPUs by default
in order to reduce database overuse, we need to restrain that on
cloudcontrol1003.  Since this is in the neutron class, we shouldn't need
to worry about changing this in main/eqiad directly for labcontrol1001.

I don't even think that actually runs the service anyway.

Bug: T188589
Change-Id: I74c02e93874267a16422ee47bfdf1fd96fc233ef