adding tonina to groups in admin module
adding user to analytics-users and researchers, per the info on the
parent task and related other wmde users getting identical access

this should only be merged by SRE team members after a review for
objeciton on the parent task.  If no objections are noted on parent
task, this can be merged on 2018-08-23.

Bug: T202069
Change-Id: I2038ea8a86c3105148cb034fa1442b57c9d7a534