Sum eventlogging throughput rate when computing quantiles for alerts
If there are multiple leaders for a topic within the given time range,
rate() will return a metric for each one.  We need the total sum rate
across all leader brokers.

NOTE: This might result in a high rate over the last period (15m)
after a leader election, but is correct for multi partition topics
and if a leader election happened recent enough that there is a
0/second rate reported by the non-leader broker.

Change-Id: I3531f6a426f986b27d385c71c45ee65861af3df8