profile::archiva::proxy: protect first run with firewall rules
The first run of Archiva after its installation allows to create
a Admin account, that will have unlimited power on Archiva. To protect
this use case when the archiva host runs on untrusted networks (like
when it is exposed to Internet), a new option to limit port 80/443's
ferm srange is introduced.

This should ease the deployment of archiva to the new host (archiva1001)
without risking that somebody external to WMF can compromise Archiva's
security.

Bug: T192639
Change-Id: I2e317d5a43040dfa05bce3c8fb3fdc7aa26ffb42