adding user Samuel Guebo to admin module
This is 1 of 2 patchsets.  This patch adds the user and their login
info, but doesn't append them into any groups.  A second patchset will
follow to add to groups.  (If a new user is added in the same patchset
as adding to groups, the CI testing will sometimes fail due to the user
not existing when the test parses the groups.)

This has the typical 3 business day wait.  If no objections are noted on
the parent task, it can be merged live on Friday, 2018-08-24.

Bug: T202362
Change-Id: Idc212495f8d7885074d91f698a85d23728ede2f2