Deploy msearch daemon to cirrus servers
Deploys the msearch service to the fleet of cirrus elasticsearch
servers. Includes new cli options for separating the relforge
and prod daemons, and for controlling the load that the daemons
can potentially generate.

The linked depends-on patch must be merged and deployed prior
to this, or the CLI options wont exist.

Depends-On: I5dd54dd5d5e5c71807c15b57d7638349dfb4ad08
Bug: T200740
Change-Id: I70700554e61b71c1830e618c919a9467c7e1f4be