netbox: fix psql client auth for monitoring check
The Icinga check for postgresql needs replication user
permission [1], connects via IPv6 and runs on the slave
to connect to localhost and the master.

This combination meant there was no pg_hba.conf entry
matching it and it was denied permission.

Former solution used template1 as default database name
but we also want to use the netbox db (with the replication user).

[1] comments on
https://github.com/wikimedia/puppet/blob/production/modules/postgresql/manifests/slave/monitoring.pp

Bug: T185504
Change-Id: Ia84bfa1504ccd241b8b71bb590a3a2a923cab429