postgresql::slave::monitoring: make check description configurable
Make it possible to specify a non-standard description for the Icinga
check.

So for example when it the check is assigned to host netmon
we want to point out it's about netbox and not one of the other services
on netmon.

Also change descriptions of other checks so that when searching for
"netbox" on Icinga you get all related checks, HTTPS, SSL, uwsgi and
the DB check.

Bug: T185504
Change-Id: I10bbf17cca18d9579e43d7344b90ee600fa3a893