nova: scale back the database and worker usage for cloudcontrol1003
The nova-network template already uses pools of 4 per worker, and they also
only have 12 workers responding.  This has reduced the bleeding of database
connections quite a lot.  cloudcontrol1003 is a large server, but it
doesn't need to be scaled heavily beyond what the current setup has at this time
to my knowledge.  This sets it's values more in line with the nova-network
setup, and can be scaled up later if it is needed.  Leaving the value of
the overflow alone means it can consume more connections if needed per worker.

Bug: T202549
Change-Id: Icd3315030579b0c006502cb7b4ea4f43d633d187