mjolnir: Enable loggers created before logging is setup
It turns out the reason we were getting such a low amount of logging
was that any loggers created before the config was passed, most of
them via imports, were being disabled. Turn them back on.

Also increases logging level to INFO. The logging level is still
pretty low (~10/min when active) and will be useful for insight into
what is going on.

Change-Id: I8d85f4496eb6c7494a09844ec8fdce8f64440499