icinga: rename role::icinga to profile::icinga
role::icinga isn't actually a role, the role
is called 'alerting_host' and includes the icinga role
and more.

Also the motivation is that there are Hiera calls i here
that should be moved to parameters of a profile in the
follow-up change.

Step by step this should be fully converted to profiles
included in the role alerting_host.

Change-Id: I616290e5827c5474245c48a18676e819715f5d77