ssh-agent-proxy: add default values to --help
It's helpful to know what the default values are in --help's output and
not just at the banner at the top of the source file. argparse makes
this easy, so just add it.

In the spirit of keeping things DRY, also remove the help output from
the comments at the top of the code; a --help invocation is easy and
will be accurate without having to keep this section in-sync.

While editing that section, also update the copyright statements there.

Change-Id: Iced651795d5d865eede9c3392f4c63874a45fd80