tor::relay: add configurable thirdparty APT source
per T196701#4537442 quoting Moritz:

<quote>
We use the debs provided by the Tor project,
not the ones from Debian.

In jessie this was in the thirdparty component which was
present by default.

On stretch the thirdparty/tor component needs to be
explicitly added via puppet.
</quote>

Doing this and making the APT source URI and dist
configurable as parameters.

To avoid hardcoded wikimedia strings inside the
module and keep it a reusable module for others.

Bug: T196701
Change-Id: I4af3feffb0497cf46eda8e02a11edbbd641107e4